refactor: migrate to python3 and retire centos pipeline
