--- centos/Dockerfile
@@ -1,27 +1,0 @@
-# Copyright 2019 Tetrate
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-
-FROM centos:7
-
-ENV PATH=/opt/rh/rh-git218/root/bin:/opt/rh/devtoolset-7/root/usr/bin:/usr/local/sbin:/usr/local/bin:/usr/sbin:/usr/bin:/sbin:/bin
-
-COPY build_container/ /build_container
-
-WORKDIR /build_container
-RUN ./getenvoy_centos.sh
-
-COPY envoy_pkg /envoy_pkg
-WORKDIR /envoy_pkg
-
-VOLUME ["/tmp/packaged"]
